integration_tests/client-library/python/requirements.txt
psycopg2-binary
pytest
sqlalchemy-risingwave
SQLAlchemy==1.4.51
